marketing-seo-website/python-seo-scanner/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0
google-ads>=24.0.0
openai>=1.0.0
markdown>=3.5.0
python-dotenv>=1.0.0
typer>=0.9.0
pydantic>=2.0.0
lxml>=4.9.0
rich>=13.0.0
